Aether_docker/IngestionService/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
confluent-kafka==2.11.1
dataclasses-json==0.6.7
deprecation==2.1.0
easyocr==1.7.2
fastapi==0.116.1
filelock==3.19.1
frozenlist==1.7.0
fsspec==2025.9.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.76
langchain-ollama==0.3.8
langchain-text-splitters==0.3.11
langgraph==0.6.7
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.6
langsmith==0.4.27
lazy_loader==0.4
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
networkx==3.5
ninja==1.13.0
numpy==2.2.6
ollama==0.5.3
opencv-python-headless==4.12.0.88
orjson==3.11.3
ormsgpack==1.10.0
packaging==25.0
pillow==11.3.0
postgrest==1.1.1
propcache==0.3.2
pyclipper==1.3.0.post6
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.10.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
realtime==2.7.0
requests==2.32.5
requests-toolbelt==1.0.0
scikit-image==0.25.2
scipy==1.16.2
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
storage3==0.12.1
StrEnum==0.4.15
supabase==2.18.1
supabase_auth==2.12.3
supabase_functions==0.10.1
sympy==1.14.0
tenacity==9.1.2
tifffile==2025.9.9
torch==2.8.0
torchvision==0.23.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
xxhash==3.5.0
yarl==1.20.1
zstandard==0.24.0
fastapi
uvicorn[standard]
python-multipart
httpx
storage3
# Add your other dependencies here